feat(comparateur): éditorialisation des réponses
